// api/models/update-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Social Media API
 * The Social Media API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface UpdateUserDTO
 */
export interface UpdateUserDTO {

    /**
     * @type {string}
     * @memberof UpdateUserDTO
     */
    name?: string;

    /**
     * @type {string}
     * @memberof UpdateUserDTO
     */
    username?: string;

    /**
     * @type {string}
     * @memberof UpdateUserDTO
     */
    about?: string;

    /**
     * @type {string}
     * @memberof UpdateUserDTO
     */
    password?: string;

    /**
     * @type {string}
     * @memberof UpdateUserDTO
     */
    oldPassword?: string;
}